@@ -3,13 +3,14 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   # Firebase
+  firebase_core: ^0.4.0+1
   firebase_remote_config: ^0.2.0+7
   firebase_analytics: ^5.0.2
   
