@@ -4,6 +4,10 @@ import 'package:nutrition_questionnaire/logic/main_bloc.dart';
 
 class StartPage extends StatelessWidget {
 	
+	final bool showStartButton;
+
+	StartPage({@required this.showStartButton});
+
     @override
     Widget build(BuildContext context) {
         return Container(
@@ -16,11 +20,15 @@ class StartPage extends StatelessWidget {
 
                         Text("Welcome to the Nutrition questionnaire"),
 
-                        MaterialButton(
-                            child: Text("Start"),
-                            onPressed: MainBloc.of(context).startQuestions,
-                        )
-
+                        AnimatedOpacity(
+							opacity: showStartButton ? 1.0 : 0.0,
+							duration: Duration(milliseconds: 300),
+							curve: Curves.easeInOut,
+							child: MaterialButton(
+                            	child: Text("Start"),
+                            	onPressed: MainBloc.of(context).startQuestions,
+                        	)
+						)
                     ],
                 ),
             ),
